# perf(ext/web): optimize performance.measure()